battery-Data_api/requirements.txt
alembic==1.13.1
aniso8601==9.0.1
blinker==1.7.0
cffi==1.16.0
click==8.1.7
cryptography==42.0.1
Flask==3.0.1
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.0
MarkupSafe==2.1.4
pycparser==2.21
PyMySQL==1.1.0
python-dotenv==1.0.1
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
Werkzeug==3.0.1
